NameSort and PriceSort added

+++ lib/product.dart
@@ -1,9 +1,8 @@
-import 'package:flutter/material.dart';
 
 class Product {
   String name;
   String imageUrl;
-  double price;
+  int price;
   String description;
 
   Product({required this.name, required this.imageUrl, required this.price, required this.description});
